refactor: extract net_stat metrics into their own PromEx plugin
Report listen_drops/listen_overflows as Prometheus counters rather
than gauges via reporter_options: [prometheus_type: "counter"], since
they're monotonic kernel counters read from /proc/net/netstat.
last_value is kept for storage since the value is already cumulative
from the kernel — summing it ourselves would double-count.
