+++ lib/supavisor/monitoring/prom_ex.ex
@@ -14,7 +14,7 @@ defmodule Supavisor.Monitoring.PromEx do
   alias PromEx.Plugins
   alias Supavisor.PeepStorage
   alias Supavisor.PeepStorage.PrometheusCached
-  alias Supavisor.PromEx.Plugins.{Cluster, OsMon, Tenant}
+  alias Supavisor.PromEx.Plugins.{Cluster, NetStat, OsMon, Tenant}
   alias Telemetry.Metrics
 
   defmodule Store do
@@ -66,6 +66,7 @@ defmodule Supavisor.Monitoring.PromEx do
 
       # Custom PromEx metrics plugins
       {OsMon, poll_rate: poll_rate},
+      {NetStat, poll_rate: poll_rate},
       {Tenant, poll_rate: poll_rate},
       {Cluster, poll_rate: poll_rate}
     ]
